+++ SupplyAndDemand/Manufacturer.cpp
@@ -9,7 +9,7 @@ Manufacturer::Manufacturer(ManufacturerSharedData* aSharedData) :
 	
 }
 
-void Manufacturer::Update(const float deltaTime, const float deltaHours)
+void Manufacturer::Update(const double deltaTime, const double deltaHours)
 {
 	Entity::Update(deltaTime, deltaHours);
 	if (CanProduce())
